RGB GUI
It introduces three color models that are used in computing and
graphics.
It gives you practice in writing complex functions with conditionals.
It gives you experience with using traces debug program flow.
It introduces you to the notion of attribute invariants.
It demonstrates a complex Python application that spans multiple
modules.